requirements.txt
absl-py==0.11.0
adal==1.2.5
affine==2.3.0
aiohttp==3.7.3
antlr4-python3-runtime==4.7.2
applicationinsights==0.11.9
appnope==0.1.0
argcomplete==1.12.1
argon2-cffi==20.1.0
asciitree==0.3.3
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
azure-appconfiguration==1.1.1
azure-batch==9.0.0
azure-cli==2.14.2
azure-cli-core==2.14.2
azure-cli-telemetry==1.0.6
azure-common==1.1.26
azure-core==1.9.0
azure-cosmos==3.2.0
azure-datalake-store==0.0.51
azure-functions-devops-build==0.0.22
azure-graphrbac==0.60.0
azure-keyvault==1.1.0
azure-keyvault-administration==4.0.0b2
azure-loganalytics==0.1.0
azure-mgmt-advisor==2.0.1
azure-mgmt-apimanagement==0.2.0
azure-mgmt-appconfiguration==0.6.0
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-authorization==0.61.0
azure-mgmt-batch==9.0.0
azure-mgmt-batchai==2.0.0
azure-mgmt-billing==0.2.0
azure-mgmt-botservice==0.2.0
azure-mgmt-cdn==5.0.0
azure-mgmt-cognitiveservices==6.3.0
azure-mgmt-compute==14.0.0
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==1.5.0
azure-mgmt-containerregistry==3.0.0rc15
azure-mgmt-containerservice==9.4.0
azure-mgmt-core==1.2.0
azure-mgmt-cosmosdb==1.0.0
azure-mgmt-datalake-analytics==0.2.1
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==0.1.0
azure-mgmt-deploymentmanager==0.2.0
azure-mgmt-devtestlabs==4.0.0
azure-mgmt-dns==2.1.0
azure-mgmt-eventgrid==3.0.0rc7
azure-mgmt-eventhub==4.1.0
azure-mgmt-hdinsight==1.7.0
azure-mgmt-imagebuilder==0.4.0
azure-mgmt-iotcentral==3.0.0
azure-mgmt-iothub==0.12.0
azure-mgmt-iothubprovisioningservices==0.2.0
azure-mgmt-keyvault==7.0.0b3
azure-mgmt-kusto==0.3.0
azure-mgmt-loganalytics==0.7.0
azure-mgmt-managedservices==1.0.0
azure-mgmt-managementgroups==0.2.0
azure-mgmt-maps==0.1.0
azure-mgmt-marketplaceordering==0.2.1
azure-mgmt-media==2.2.0
azure-mgmt-monitor==0.11.0
azure-mgmt-msi==0.2.0
azure-mgmt-netapp==0.12.0
azure-mgmt-network==12.0.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.5.0
azure-mgmt-privatedns==0.1.0
azure-mgmt-rdbms==3.1.0rc1
azure-mgmt-recoveryservices==0.4.0
azure-mgmt-recoveryservicesbackup==0.6.0
azure-mgmt-redhatopenshift==0.1.0
azure-mgmt-redis==7.0.0rc2
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.6.0
azure-mgmt-resource==10.2.0
azure-mgmt-search==2.1.0
azure-mgmt-security==0.4.1
azure-mgmt-servicebus==0.6.0
azure-mgmt-servicefabric==0.5.0
azure-mgmt-signalr==0.4.0
azure-mgmt-sql==0.21.0
azure-mgmt-sqlvirtualmachine==0.5.0
azure-mgmt-storage==11.2.0
azure-mgmt-synapse==0.3.0
azure-mgmt-trafficmanager==0.51.0
azure-mgmt-web==0.48.0
azure-multiapi-storage==0.5.2
azure-nspkg==3.0.2
azure-storage-blob==12.6.0
azure-storage-common==1.4.2
azure-synapse-accesscontrol==0.2.0
azure-synapse-artifacts==0.3.0
azure-synapse-spark==0.2.0
backcall==0.2.0
bcrypt==3.2.0
bleach==3.2.1
bokeh==2.2.3
boto==2.49.0
branca==0.4.1
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.3
cftime==1.2.1
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.7.0
cloudpickle==1.6.0
colorama==0.4.4
crcmod==1.7
cryptography==2.9.2
cycler==0.10.0
dask==2.30.0
decorator==4.4.2
defusedxml==0.6.0
descartes==1.1.0
distributed==2.30.1
earthengine-api==0.1.242
entrypoints==0.3
fabric==2.5.0
fasteners==0.15
Fiona==1.8.17
folium==0.11.0
fsspec==0.8.4
future==0.18.2
gast==0.3.3
gcs-oauth2-boto-plugin==2.7
geographiclib==1.50
geopandas==0.8.1
geopy==2.0.0
google-api-core==1.23.0
google-api-python-client==1.12.5
google-apitools==0.5.31
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-cloud-core==1.4.3
google-cloud-storage==1.32.0
google-crc32c==1.0.0
google-pasta==0.2.0
google-reauth==0.1.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
gsutil==4.55
h5netcdf==0.8.1
h5py==3.1.0
HeapDict==1.0.1
httplib2==0.18.1
httplib2shim==0.0.3
humanfriendly==8.2
idna==2.10
invoke==1.4.1
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
isodate==0.6.0
javaproperties==0.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
jsmin==2.2.2
json5==0.9.5
jsondiff==1.2.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
knack==0.7.2
locket==0.2.0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mock==2.0.0
monotonic==1.5
msal==1.0.0
msal-extensions==0.1.3
msgpack==1.0.0
msrest==0.6.19
msrestazure==0.6.4
munch==2.5.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
netCDF4==1.5.4
notebook==6.1.5
numcodecs==0.7.2
numpy==1.19.4
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.7.1
partd==1.1.0
pbr==5.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pkginfo==1.6.1
portalocker==1.7.1
progressbar==2.5
progressbar2==3.53.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.13.0
psutil==5.7.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.2
PyJWT==1.7.1
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyproj==3.0.0.post1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2019.1
pyu2f==0.1.5
PyYAML==5.3.1
pyzmq==20.0.0
rasterio==1.1.8
rasterstats==0.15.0
requests==2.25.0
requests-oauthlib==1.3.0
retry-decorator==1.1.1
rsa==4.6
Rtree==0.9.4
scipy==1.5.4
scp==0.13.3
seaborn==0.11.0
Send2Trash==1.5.0
Shapely==1.7.1
simplejson==3.17.2
six==1.15.0
snuggs==1.4.7
sortedcontainers==2.3.0
sshtunnel==0.1.5
tabulate==0.8.7
tblib==1.7.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
toolz==0.11.1
tornado==6.1
tqdm==4.51.0
traitlets==5.0.5
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
utm==0.6.0
vsts==0.1.25
vsts-cd-manager==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==1.0.1
wget==3.2
wrapt==1.12.1
xarray==0.16.1
xmltodict==0.12.0
zarr==2.5.0
zict==2.0.0
